requirements.txt
fastapi==0.111.0
mangum==0.17.0
pymongo==4.7.2
python-dotenv==1.0.1
uvicorn==0.30.1
jinja2==3.1.4